+++ packages/bot/src/commands/config/muteRole.ts
@@ -4,7 +4,7 @@ import { MESSAGES } from '../../util/Constants';
 
 export default class MuteRoleCommand extends Command {
   public constructor() {
-    super('muteRole', {
+    super('set-muteRole', {
       category: 'config',
       args: [
         {
@@ -47,6 +47,8 @@ export default class MuteRoleCommand extends Command {
 
           case 'voice':
             overwrite.CONNECT = false;
+            break;
+          default:
             break;
         }