+++ packages/bot/src/struct/Redis.ts
@@ -1,7 +1,7 @@
 import redisClient = require('ioredis');
 
 export default new redisClient({
-  host: process.env.REDIS_HOST || 'localhost',
-  db: Number(process.env.REDIS_DB) || 0,
+  host: process.env.REDIS_HOST ?? 'localhost',
+  db: Number(process.env.REDIS_DB),
   password: process.env.REDIS_PASSWORD
 });