+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next build 21 (0.2.0) (2020-05-19)
+# Next build 28 (0.2.0) (2020-05-24)
 
 ## Fixes
 
@@ -10,4 +10,13 @@
 - actually exit process if on unhandled rejection if in prod ([b294f14](https://github.com/weeb-cafe/MewChan/commit/b294f14a1ffdef11f763a1680f2d23dfe0b0cb91))
 
 ## Chores
+
+### Repo
+- update @typescript-eslint to v3 & eslint-config-marine to v7
 - various major package updates ([6885ba2](https://github.com/weeb-cafe/MewChan/commit/6885ba2cb9855ba685bba5d95ba951e29748f9dc), [31f1916](https://github.com/weeb-cafe/MewChan/commit/31f19165697f6485038d9d8d0ac5fe49d88be2bc), [0615295](https://github.com/weeb-cafe/MewChan/commit/0615295cf5a80973509c4efa28bf278cd1da577c))
+
+## Features
+
+### Bot
+- reaction roles ([eae5869](https://github.com/weeb-cafe/MewChan/commit/eae586977a1d3bb1c581eeb3bcb4303cdf1beb44))
+- help command ([0de4fd9](https://github.com/weeb-cafe/MewChan/pull/24/commits/0de4fd92091ed51c18a558549d8be2bd02ee1d92))